requirements.txt
asgiref==3.4.1
attrs==21.2.0
bleach==4.1.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-summernote==0.8.20.0
djangorestframework==3.12.4
drf-spectacular==0.21.0
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.2.1
Markdown==3.3.4
packaging==21.2
Pillow==8.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
webencodings==0.5.1
zipp==3.6.0
